webhook_receiver/requirements.txt
asana==5.2.4
fastapi==0.128.8
uvicorn[standard]==0.39.0
httpx==0.28.1
google-cloud-secret-manager==2.27.0
google-cloud-logging==3.15.0
python-dotenv==1.2.1
